# docs/jiuyangongshe_openapi.yaml
openapi: 3.0.3
info:
  title: 韭研公社 API
  description: |
    韭研公社 APP API 文档（基于 v1.3.4 版本抓包逆向）

    ## 认证说明
    - 登录接口无需认证
    - 其他接口需要在 Header 中传入 `token`（登录返回的 sessionToken）和 `Cookie: SESSION=<sessionToken>`
    - 设备 token（用于首次登录）：`63C8320CE1786D66911ED9D79F96D9E7`

    ## 通用 Headers
    ```
    Content-Type: application/json
    version_name: 1.3.4
    version: 2025070723
    platform: 1
    token: <sessionToken>
    timestamp: <毫秒时间戳>
    ```
  version: 1.0.0
  contact:
    name: Daily Stock Analysis
servers:
  - url: https://app.jiuyangongshe.com/jystock-app/api
    description: 韭研公社 APP API 服务器

paths:
  /v1/user/login:
    post:
      tags: [认证]
      summary: 用户登录
      description: 使用手机号和密码登录，返回 sessionToken 和 SESSION cookie。设备 token 会自动注入。
      operationId: userLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              phone: "15160047541"
              password: "Am112211"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "9":
          description: 版本过低

  /v1/user/info:
    post:
      tags: [认证]
      summary: 获取用户信息
      operationId: getUserInfo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        "1":
          description: 登录失效

  /v1/user/notice/broadcast:
    post:
      tags: [内容列表]
      summary: 获取广播消息列表（盘前纪要/涨停简图/晚间公告等）
      description: |
        返回关注的广播消息列表，包含盘前纪要、涨停简图、晚间公告、内容精选等。
        返回的 `target_id` 可直接用于 `/v2/article/detail` 获取文章详情。
      operationId: getBroadcastList
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BroadcastListRequest'
            example:
              start: "1"
              limit: "10"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BroadcastListResponse'
        "9":
          description: 服务繁忙

  /v2/article/detail:
    post:
      tags: [内容列表]
      summary: 获取文章详情
      description: |
        根据 `article_id`（或 broadcast 接口返回的 `target_id`）获取文章完整内容。
        返回 HTML 格式内容，需自行去除标签。
      operationId: getArticleDetail
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleDetailRequest'
            example:
              is_read: "1"
              article_id: "4efqr8fdln2"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleDetail'

  /v2/user/article/follow:
    post:
      tags: [内容列表]
      summary: 获取关注者文章列表
      description: |
        按 type 分类获取用户关注的文章列表。
        type 可选值：1=综合, 2=题材, 3=公告, 4=我的自选, 5=其他
      operationId: getFollowArticles
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowArticlesRequest'
            example:
              type: "1"
              start: "1"
              limit: "10"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleListResponse'

  /v1/industry/list:
    post:
      tags: [题材数据]
      summary: 获取题材库列表
      description: 获取热门题材列表，如磷化铟、宁德时代概念、光通信等。
      operationId: getIndustryList
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndustryListRequest'
            example:
              keyword: ""
              limit: "15"
              start: "1"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndustryListResponse'

  /v1/action/diagram-url:
    post:
      tags: [涨停数据]
      summary: 获取涨停简图 URL
      description: 根据日期获取当天涨停涨停简图的图片 URL。
      operationId: getDiagramUrl
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiagramUrlRequest'
            example:
              date: "2026-04-17"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagramUrlResponse'

  /v1/action/list:
    post:
      tags: [涨停数据]
      summary: 获取涨停股列表
      description: |
        获取涨停股详情列表，按日期和板块分类。
        `action_field_id=all,YYYY-MM-DD` 获取当日全部涨停
        `action_field_id=recommend,YYYY-MM-DD` 获取当日推荐涨停
      operationId: getActionList
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionListRequest'
            example:
              action_field_id: "all,2026-04-17"
              start: "1"
              limit: "10"
              sort_time: "0"
              sort_price: "0"
              sort_range: "0"
              is_filter_st: "0"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionListResponse'

  /v1/product/article/list:
    post:
      tags: [内容列表]
      summary: 获取产品文章列表（如逻辑红宝书）
      description: 获取韭研公社产品文章列表，如"逻辑红宝书每日市场精选"。
      operationId: getProductArticles
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductArticleListRequest'
            example:
              product_id: "1"
              start: "1"
              limit: "15"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductArticleListResponse'

  /v1/timeline/list:
    post:
      tags: [大事件日历]
      summary: 获取大事件日历
      description: 按年月获取重要财经事件时间线。
      operationId: getTimeline
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimelineRequest'
            example:
              date: "2026-04"
              grade: "0"
              keyword: ""
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineResponse'

  /v1/user/notice:
    post:
      tags: [系统消息]
      summary: 获取用户通知消息
      description: 获取系统通知、收益信息等用户消息。
      operationId: getUserNotice
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserNoticeResponse'

components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: token
      description: 登录返回的 sessionToken

  schemas:
    LoginRequest:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
          description: 手机号
          example: "15160047541"
        password:
          type: string
          description: 密码
          example: "Am112211"

    LoginResponse:
      type: object
      properties:
        msg:
          type: string
        errCode:
          type: string
          enum: ["0", "9"]
        data:
          type: object
          properties:
            user_id:
              type: string
            phone:
              type: string
            nickname:
              type: string
            sessionToken:
              type: string
              description: 后续请求需在 Header 中传入此 token
            avatar:
              type: string
            follow_count:
              type: integer
            fans_count:
              type: integer
        serverTime:
          type: integer

    UserInfo:
      type: object
      properties:
        msg:
          type: string
        errCode:
          type: string
        data:
          type: object
          properties:
            user_id:
              type: string
            phone:
              type: string
            nickname:
              type: string
            avatar:
              type: string
            follow_count:
              type: integer
            fans_count:
              type: integer

    BroadcastListRequest:
      type: object
      required: [start, limit]
      properties:
        start:
          type: string
          description: 分页起始
          example: "1"
        limit:
          type: string
          description: 每页数量
          example: "10"

    BroadcastListResponse:
      type: object
      properties:
        msg:
          type: string
        errCode:
          type: string
        data:
          type: object
          properties:
            pageNo:
              type: integer
            pageSize:
              type: integer
            result:
              type: array
              items:
                $ref: '#/components/schemas/BroadcastItem'

    BroadcastItem:
      type: object
      properties:
        title:
          type: string
          description: 文章标题
          example: "4月17日盘前纪要"
        create_time:
          type: string
          description: 创建时间
          example: "2026-04-17 07:31:43"
        target_id:
          type: string
          description: 文章 ID（用于调 article/detail）
          example: "2z1p49jw0wy"
        type:
          type: integer
          description: 类型标识

    ArticleDetailRequest:
      type: object
      required: [is_read, article_id]
      properties:
        is_read:
          type: string
          description: 是否标记已读
          example: "1"
        article_id:
          type: string
          description: 文章 ID（broadcast 返回的 target_id）
          example: "2z1p49jw0wy"

    ArticleDetail:
      type: object
      properties:
        msg:
          type: string
        errCode:
          type: string
        data:
          type: object
          properties:
            article_id:
              type: string
            title:
              type: string
            content:
              type: string
              description: HTML 格式文章内容
            create_time:
              type: string
            comment_count:
              type: integer
            like_count:
              type: integer
            user:
              type: object
              properties:
                nickname:
                  type: string
            stock_list:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                    example: "宁德时代"
                  code:
                    type: string
                    example: "sz300750"

    FollowArticlesRequest:
      type: object
      required: [type, start, limit]
      properties:
        type:
          type: string
          enum: ["1", "2", "3", "4", "5"]
          description: "1=综合, 2=题材, 3=公告, 4=我的自选, 5=其他"
        start:
          type: string
        limit:
          type: string

    ArticleListResponse:
      type: object
      properties:
        msg:
          type: string
        errCode:
          type: string
        data:
          type: object
          properties:
            pageNo:
              type: integer
            pageSize:
              type: integer
            result:
              type: array
              items:
                type: object
                properties:
                  article_id:
                    type: string
                  title:
                    type: string
                  create_time:
                    type: string
                  stock_list:
                    type: array
                    items:
                      type: object

    IndustryListRequest:
      type: object
      required: [keyword, limit, start]
      properties:
        keyword:
          type: string
        limit:
          type: string
        start:
          type: string

    IndustryListResponse:
      type: object
      properties:
        msg:
          type: string
        errCode:
          type: string
        data:
          type: object
          properties:
            result:
              type: array
              items:
                type: object
                properties:
                  industry_id:
                    type: string
                  title:
                    type: string
                  keyword:
                    type: string

    DiagramUrlRequest:
      type: object
      required: [date]
      properties:
        date:
          type: string
          description: 日期格式 YYYY-MM-DD
          example: "2026-04-17"

    DiagramUrlResponse:
      type: object
      properties:
        msg:
          type: string
        errCode:
          type: string
        data:
          type: string
          format: uri
          description: 涨停简图图片 URL

    ActionListRequest:
      type: object
      required: [action_field_id, start, limit, sort_time, sort_price, sort_range, is_filter_st]
      properties:
        action_field_id:
          type: string
          description: "板块ID，格式如 all,YYYY-MM-DD（全部）或 recommend,YYYY-MM-DD（推荐）"
          example: "all,2026-04-17"
        start:
          type: string
        limit:
          type: string
        sort_time:
          type: string
        sort_price:
          type: string
        sort_range:
          type: string
        is_filter_st:
          type: string

    ActionListResponse:
      type: object
      properties:
        msg:
          type: string
        errCode:
          type: string
        data:
          type: object
          properties:
            result:
              type: array
              items:
                type: object
                properties:
                  code:
                    type: string
                    example: "sz002528"
                  name:
                    type: string
                    example: "ST英飞拓"
                  article:
                    type: object

    ProductArticleListRequest:
      type: object
      required: [product_id, start, limit]
      properties:
        product_id:
          type: string
          description: 产品 ID，如 "1" 为逻辑红宝书
          example: "1"
        start:
          type: string
        limit:
          type: string

    ProductArticleListResponse:
      type: object
      properties:
        msg:
          type: string
        errCode:
          type: string
        data:
          type: object
          properties:
            result:
              type: array
              items:
                type: object
                properties:
                  product_article_id:
                    type: string
                  title:
                    type: string
                  describe:
                    type: string
                    description: 文章摘要
                  create_time:
                    type: string
                  author:
                    type: object

    TimelineRequest:
      type: object
      required: [date, grade, keyword]
      properties:
        date:
          type: string
          description: 年份或年月，如 "2026" 或 "2026-04"
          example: "2026-04"
        grade:
          type: string
        keyword:
          type: string

    TimelineResponse:
      type: object
      properties:
        msg:
          type: string
        errCode:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              list:
                type: array

    UserNoticeResponse:
      type: object
      properties:
        msg:
          type: string
        errCode:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              type:
                type: integer
              title:
                type: string
              content:
                type: string
              create_time:
                type: string